+++ CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 - added erraux.Causer and the ability to customize cause in erraux.Encoder
+- changed the signature of ErrorFielder.ErrorFields to avoid a dependency on erraux
 
 ## [v0.1.6]
 - allow appending and extending Header while mainting immutability